Allow translation of ObjectIntersectionOf-type class expressions.
When creating a composite cross-species ontology, translate class
expressions of type ObjectIntersectionOf in addition to expressions of
type ObjectSomeValuesOf.

Since the original OWLTools implementation didn't do that, hide that
feature behind an option.

Of note, recent versions of FBbt contain class expressions that require
such a feature.

Whether the feature is enabled or disabled, always check whether an
expression has been translated successfully before attempting to craft
an axiom with the translated expression. Failure to do so may result in
a crash when processing an unsupported class expression (as it happens
with FBbt, when translation of ObjectIntersectionOf expressions is
disabled).