Move all classes into 'org.incenp.obofoundry.uberon' package.
This is no longer a demonstration project for ROBOT plugins, but a set
of plugins intended to be used in the Uberon pipeline. So rename the
package accordingly.

In the process, we put all classes in a single package, instead of one
package for the ROBOT commands and one package for the "helper" classes.
There are not enough classes to justify separating them in two distinct
packages. KISS.