Build a standalone version of ROBOT.
As the upstream version of ROBOT has no support for pluggable commands
yet, we include our own main entry point that behaves as the original
entry point but with our two extra commands included. We build a
"shaded" Jar file that includes all of the original ROBOT, but with our
own entry point as the main class.