Remove declaration axioms for merged classes.
When a class has been merged into another as a result of a "merge
species" operation, the class declaration axiom for that class is no
longer needed and can be dropped.

This commit adds an option to the "merge-species" command to do just
that. It is an optional behaviour because it differs from the behaviour
of the original command in OWLTools.