Do not exclude an axiom for lack of an ObjectProperty in signature.
When performing a "species merge", if the -q option has been used, make
sure to only exclude an axiom if it is referring to an object property
and that property is not one of those that must be preserved. If the
axiom is not referring to any object property at all in its signature,
then accept the axiom as if the -q option has not been used.