Add options to deal with general class axioms.
Current implementation of the "species merge" leaves general class
axioms as they are, even if they refer to a class that has been merged
into another one. This results both in the merged ontology still having
references to merged classes, and in the general class axioms losing
their actual meaning.

Here we add two options to the "merge-species" command to deal with
general class axioms that refer to merged classes: with "-g", such
axioms will be translated if possible (or dropped if they can't be
translated), and with "-G", they will be forcibly dropped without any
attempt at translating them.