Translate more class expressions when using -x option.
Allow the translation of more class expressions than just
OWLObjectSomeValuesFrom and OWLObjectIntersectionOf: we also translate:

* OWLObjectComplementOf,
* OWLObjectUnionOf,
* OWLObjectExactCardinality,
* OWLObjectMinCardinality, and
* OWLObjectMaxCardinality.

By default, only OWLObjectSomeValuesFrom expressions are translated.
Translation of other types of expression must be enabled by the -x
option, whose meaning is extended to indicate "translation of everything
that was not translated in the original OWLTools command".