Allow choosing the strategy for species subsetting.
Implement a different strategy for species subsetting: instead of
asserting "owl:Thing in_taxon some <the_taxon>" and excluding all
unsatisfiable classes, we include each class for which the assertion
"<the_class> and in_taxon some <the_taxon>" is satisfiable (as suggested
in the comments of OWLTools' SpeciesSubsetterUtil class).

The new strategy is ~3-4 times slower, and may include classes that
would have been excluded by the former strategy when property chains
over in_taxon are missing. For example, if we have

  C1 adjacent_to some C2

and C2 is not in the desired taxon ("C2 and in_taxon some <the_taxon>"
is unsatisfiable), "C1 and in_taxon some <the_taxon>" might still be
satisfiable because there is no property chain between adjacent_to and
in_taxon. Using the former strategy, C1 would automatically be
unsatisfiable simply because it depends on C2 which is unsatisfiable.

On the other hand, the former strategy may lead to the exclusion of
classes that should not be excluded, as soon as they are linked to any
unsatisfiable classes (regardless of the existence of any property chain
over in_taxon).

Both strategies have their shortcomings, so... we enable both, and let
the user decide.