Do not exclude classes linked by "cross-taxon relations".
When using the "default" subsetting strategy (removing all unsatisfiable
classes after asserting "owl:Thing in_taxon some <the_taxon>"), some
classes may be excluded because they are linked to an unsatisfiable
class even though the link does not convey the meaning of "in taxon".

For example, if we have

  C1 shares_ancestor_with C2

and C2 is unsatisfiable in the taxon we are subsetting for, then C1
may be excluded solely because it depends on an unsatisfiable class,
even though it may actually be a valid class for the considered taxon.

To prevent this, we forcefully remove some "cross-taxon relationships"
from the ontology before subsetting.