Tweak command line options of `create-species-subset` command.
Change the default behaviour of the `create-species-subset` command to
remove the not-in-subset classes from the ontology by default, unless
the `--no-remove` option is specified.

Also allow to use a CURIE rather than a full IRI for the `--subset-name`
option.

Document the command, along with those changes, in the README file.